// source/frontend/src/models/api/generated/ApiGen_Concepts_DispositionFileAppraisal.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDate } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_BaseConcurrent } from './ApiGen_Base_BaseConcurrent';

// LINK: @backend/apimodels/Models/Concepts/DispositionFile/DispositionFileAppraisalModel.cs
export interface ApiGen_Concepts_DispositionFileAppraisal extends ApiGen_Base_BaseConcurrent {
  id: number | null;
  dispositionFileId: number;
  appraisedAmount: number | null;
  appraisalDate: UtcIsoDate | null;
  bcaValueAmount: number | null;
  bcaRollYear: string | null;
  listPriceAmount: number | null;
}